use keycloak for user management
don't bother with our own user management, just listen to keycloak.

whereis.auth loads up a keycloak configuration and can communicate with keycloak when given an access_token blob. These methods don't work if you give it the whole token blob (only with the access_token). Nothing is wired up to requests, either, so this is a pretty isolated setup.

Removed the jwt code, and cleaned up the middleware a bit. I think the keycloak.json pulled in as a resource should be brought in via env config since it has a secret in it.